pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acunetix-mcp-server"
version = "1.0.0"
description = "Full-coverage MCP Server for Acunetix Scanner API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
acunetix-mcp-server = "acunetix_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
